# back-end/docs/api/swagger.yaml
definitions:
  apperrors.AppError:
    properties:
      http_status:
        type: integer
      message:
        type: string
    type: object
...
  model.CampaignRequest:
    properties:
      budget:
        type: number
      goal_percentage_change:
        minimum: 0
        type: number
      max_time_between_transactions:
        minimum: 0
        type: integer
      max_transactions_amount:
        type: number
      min_time_between_transactions:
        minimum: 0
        type: integer
      min_transactions_amount:
        type: number
      parallel_transactions_amount:
        type: integer
      slippage:
        maximum: 100
        minimum: 0
        type: number
      transaction_speed:
        allOf:
        - $ref: '#/definitions/model.TransactionSpeed'
        enum:
        - default
        - fast
        - extra
    type: object
  model.CampaignSummary:
    properties:
      budget:
        type: number
      campaign_id:
        type: string
      project_id:
        type: integer
      spent_budget:
        type: number
      status:
        type: string
      token_mint_from:
        type: string
      token_mint_to:
        type: string
      type_name:
        type: string
    type: object
  model.CampaignType:
    properties:
      name:
        type: string
    type: object
...
  model.SwapCampaign:
    properties:
      budget:
        type: number
      created_at:
        type: string
      current_price:
        type: string
      goal_percent_change:
        type: number
      goal_price:
        type: string
      id:
        type: string
      max_time_between_transactions:
        type: integer
      max_transactions_budget:
        type: number
      min_time_between_transactions:
        type: integer
      min_transactions_budget:
        type: number
      parallel_transactions_amount:
        type: integer
      pool_id:
        type: string
      project_id:
        type: integer
      slippage_bps:
        type: integer
      started_price:
        type: string
      status:
        type: string
      token_mint_from:
        type: string
      token_mint_to:
        type: string
      transaction_speed:
        $ref: '#/definitions/model.TransactionSpeed'
      type:
        $ref: '#/definitions/model.CampaignType'
      updated_at:
        type: string
      using_jito:
        type: boolean
    type: object
  model.SwapTransaction:
    properties:
      address_from:
        type: string
      address_to:
        type: string
      amount_token_from:
        type: string
      amount_token_to:
        type: string
      campaign_id:
        type: string
      created_at:
        type: string
      id:
        type: integer
      message:
        type: string
      pool_id:
        type: string
      status:
        type: string
      token_mint_from:
        type: string
      token_mint_to:
        type: string
      transaction_hash:
        type: string
    type: object
...
  model.TransactionSpeed:
    enum:
    - default
    - fast
    - extra
    type: string
    x-enum-varnames:
    - Default
    - Fast
    - Extra
...
host: https://api-stage.tsunammi.io/
info:
  contact: {}
  description: This is the swagger specification for Tsunammi API.
  title: Tsunammi API
  version: 0.0.7
paths:
...
  /campaigns:
    get:
      description: Returns all campaigns associated with the authenticated user.
      operationId: get-campaigns
      parameters:
      - description: Authentication token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.SwapCampaign'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.AppError'
      security:
      - BearerAuth: []
      summary: List campaigns
      tags:
      - campaigns
  /campaigns/{id}:
    delete:
      description: Stops an active campaign by its UUID.
      operationId: stop-campaign
      parameters:
      - description: Authentication token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Campaign UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request (invalid UUID)
          schema:
            $ref: '#/definitions/apperrors.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.AppError'
      security:
      - BearerAuth: []
      summary: Stop campaign
      tags:
      - campaigns
    get:
      description: Returns a specific campaign by its UUID for the authenticated user.
      operationId: get-campaign-by-id
      parameters:
      - description: Authentication token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Campaign UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SwapCampaign'
        "400":
          description: Bad Request (invalid UUID)
          schema:
            $ref: '#/definitions/apperrors.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apperrors.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.AppError'
      security:
      - BearerAuth: []
      summary: Get campaign by ID
      tags:
      - campaigns
    put:
      consumes:
      - application/json
      description: Updates configuration for a specific campaign.
      operationId: update-campaign
      parameters:
      - description: Authentication token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Campaign UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Update parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CampaignRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrors.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.AppError'
      security:
      - BearerAuth: []
      summary: Update campaign
      tags:
      - campaigns
  /campaigns/{id}/transactions:
    get:
      description: Returns a list of transactions associated with a specific campaign.
      operationId: get-campaign-transactions
      parameters:
      - description: Authentication token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Campaign UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.SwapTransaction'
            type: array
        "400":
          description: Bad Request (invalid UUID)
          schema:
            $ref: '#/definitions/apperrors.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.AppError'
      security:
      - BearerAuth: []
      summary: Get campaign transactions
      tags:
      - campaigns
  /campaigns/summary:
    get:
      description: Returns a statistical summary of all campaigns for the authenticated
        user (e.g., total spent, active count).
      operationId: get-campaigns-summary
      parameters:
      - description: Authentication token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Campaigns summary data
          schema:
            items:
              $ref: '#/definitions/model.CampaignSummary'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.AppError'
      security:
      - BearerAuth: []
      summary: Get campaigns summary
      tags:
      - campaigns
...
securityDefinitions:
  BearerAuth:
    description: Type 'Bearer' followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"